[0003] End of Day Commit
2013-07-19
- started thinking about logics behind filters: nodes need to be able to
connect to multiple other nodes
	- started implementing del button on arrow
- finalized threaded backup_source pre-computing, exit handling, etc.;
computing now per backup_source. Each backup_source has its own
backup_ctrl.

................................
[.]		Q/A
		........................
[X]		clean-up old, obsolete code (commented out etc.)
[X]		check module-structure
[.]		maintain integrity:
		[.]	document code
			> up to backup.py, l724
		[ ]	validate user-input
		[ ]	write test/test
		[ ]	manage conditional exception-handling
		[ ]	implement correct return data for *everything*
		[ ]	address warnings/errors
		[ ] standardize logging-messages
		[ ]	make sure all logging messages are very *specific* in regards to
				- origin of message
[ ]		clean-up current code:
		[ ]	update representations
		[ ] any code-warnings/errors
		[ ]	sort methods, variables alphabetically
		[ ]	remove unused imports
		[ ]	Make sure all modified-flags (*) are reset
................................
................................
[.]		NEXT MILESTONE: 0003
		........................
[.]		views:
		[X]	basic main-window
		[X]	basic login-view
		[.]	basic set-management view
			[.]	list widget for sets
				[ ]	ability to create a new set
				[X]	ability to delete an existing set
				[X]	edit an existing set
					[X]	add an existing unused source
					[X]	add an existing unused filter
					[X] remove a source in use by current set
					[X] remove a filter in use by current set
			[.]	details widget to visualize/edit/execute etc. sets
				[.]	sources
					[X]	ability to add a source
					[X]	ability to remove a source
					[.]	ability to measure/visualize accumulated new data amount in
relation to all other sources (comparing bar)
						[(X)]	install pre-backup-sim on central location to backup_set so
it can be called from any backup_source-node and update all
backup_source nodes in set as it simulates
							> installed on each backup_source, so central location can
iterate over all backup_sources to do a batch-process
						[.]	implement graphical visualization (percentage bar relative to
the backup_source with highest accumulated data to be backed up)
							[ ]	look into Qt animations (wading throbber e.g. while
scanning~)
				[.]	filters
					[X]	ability to add a filter
					[X]	ability to remove a filter
					[ ]	add name and use as title on node ("Files > 1GiB"/".pdf Files")
					[ ]	display list of regex patterns (for now only this internal
detail; more advanced user-friendly front-end representations later)
				[.]	targets
					[ ]	ability to add a target
					[ ]	ability to remove a target
					[ ]	think about: ?ability to measure/visualize the space left on
the target?
				[ ]	Arrows
					[.] Implement support to connect multiple arrows to multiple nodes
						[.]	Del button on arrow
				[ ]	execution-panel...
		[ ]	refactor object reference names, implement consistent scheme
		[ ]	implement consistent scheme for passing object-references into
classes
[X]		controllers:
		[X]	refactor (session) controllers to follow a paradigm where "add"
and "remove" perform associative changes, while "create" and "delete" do
integral changes (to model/database)
................................
[ ]		Q/A
		........................
[ ]		.
................................
................................
[ ]		NEXT MILESTONE: 0004
		........................
[ ]		.
................................
[ ]		Q/A
		........................
[ ]		.
................................
................................
[ ]		LATER
		........................
[ ]	File-Backup
		[ ]	Implement sidecart-metadata files for backed-up datastreams
................................
[ ]		Q/A
		........................
[ ]		.
................................

+++ src/bs/gui/lib.py
@@ -1,4 +1,10 @@
 # -*- coding: utf-8 -*-
+from PySide import QtCore, QtGui
+import binascii
+import bs.config
+import bs.gui.view_sets
+import logging
+import re
 
 ###############################################################################
 ##    bs.gui.nodes                                                           ##
@@ -16,11 +22,6 @@
 ###############################################################################
 
 """ * """
-from PySide import QtCore, QtGui
-import binascii
-import bs.config
-import bs.gui.view_sets
-import re
 
 
 class BSFrame(QtGui.QFrame):
@@ -258,6 +259,7 @@ class BSArrow(QtGui.QWidget):
     _stroke_width = None
     _line_cap = None
     _join_style = None
+    _btn_del = None
 
     def __init__(self, source, target):
         super(BSArrow, self).__init__(source.parent())
@@ -277,6 +279,8 @@ class BSArrow(QtGui.QWidget):
         # INIT UI
         self.lower()
         self.show()
+        # Delete button
+        self._btn_del = BSArrowBtnDel(self)
 
     @property
     def source(self):
@@ -345,6 +349,12 @@ class BSArrow(QtGui.QWidget):
                          self._join_style)
         painter.setPen(pen)
         painter.drawPath(path)
+        # del button
+        side_len = 12
+        self._btn_del.setGeometry(p1a.x() + (p2a.x() - p1a.x()) / 2 - side_len / 2,
+                                  p1a.y() + (p2a.y() - p1a.y()) / 2 - side_len / 2,
+                                  side_len,
+                                  side_len)
 
         super(BSArrow, self).paintEvent(e)
 
@@ -365,6 +375,38 @@ class BSArrow(QtGui.QWidget):
                          self._target.geometry().united(self._source.geometry()).width() + margin * 2,
                          self._target.geometry().united(self._source.geometry()).height() + margin * 2
                          )
+
+
+class BSArrowBtnDel(BSFrame):
+    """ * """
+    _bs_arrow = None
+
+    _mouse_press_global_pos = None
+
+    def __init__(self, bs_arrow):
+        super(BSArrowBtnDel, self).__init__(bs_arrow)
+
+        self._bs_arrow = bs_arrow
+
+        self._init_ui()
+
+    def _init_ui(self):
+        css = "BSFrame {background: #C7C7FF; border-radius: 6px}"
+        css += "BSFrame:hover {background: #FF0000}"
+        self.setStyleSheet(css)
+        self.show()
+        self.raise_()
+
+    def mouseMoveEvent(self, e):
+        """ * """
+
+    def mousePressEvent(self, e):
+        self._mouse_press_global_pos = e.globalPos()
+
+    def mouseReleaseEvent(self, e):
+        if self._mouse_press_global_pos == e.globalPos():
+            if e.button() & QtCore.Qt.MouseButton.LeftButton:
+                print("Del!!!")
 
 
 class BSArrowCarrier(BSDraggable):
@@ -601,6 +643,7 @@ class BSNode(BSDraggable):
         self._layout.addWidget(self._custom_contents_container, 1, 0, 1, 1)
         self._layout.setSpacing(1)
         self._layout.setContentsMargins(0, 0, 0, 0)
+        self.setMinimumWidth(170)
         # Drop shadow
         gfx = QtGui.QGraphicsDropShadowEffect(self)
         gfx.setOffset(0)
@@ -700,13 +743,24 @@ class BSNode(BSDraggable):
         Removes the node's arrows. Additional operations take place on
         inheriting objects.
         """
-        # delete arrows
-        while len(self._arrows_inbound) > 0:
-            arrow_inbound = self._arrows_inbound[0]
-            arrow_inbound.delete()
-        if self._arrow_outbound:
-            self._arrow_outbound.delete()
-        self._bs.set_modified()
+        # request exits
+        if self.request_exit():
+            # delete arrows
+            while len(self._arrows_inbound) > 0:
+                arrow_inbound = self._arrows_inbound[0]
+                arrow_inbound.delete()
+            if self._arrow_outbound:
+                self._arrow_outbound.delete()
+            self._bs.set_modified()
+        else:
+            logging.warning("%s: Node could net be removed as active threads"\
+                            "did not shut down properly." % (self.__class__.__name__, ))
+
+    def request_exit(self):
+        """ *
+        To be overridden by inheriting objects.
+        """
+        return True
 
 
 class BSNodeConnPad(QtGui.QFrame):
@@ -728,9 +782,21 @@ class BSNodeConnPad(QtGui.QFrame):
         icon = QtGui.QFrame()
         css = "QFrame {border-top-right-radius: 2px; border-bottom-right-radius: 2px;  background: #%s}" % (bs.config.PALETTE[1], )
         css += "QFrame:hover {background: #%s}" % (bs.config.PALETTE[4], )
-        icon.setStyleSheet(css)
-        icon.setMinimumSize(QtCore.QSize(14, 14))
-        self._layout.addWidget(icon, 0, 0, 1, 1)
+        self.setMaximumWidth(14)
+        self.setMinimumWidth(14)
+        self.setStyleSheet(css)
+#         icon.setMinimumSize(QtCore.QSize(14, 14))
+#         self._layout.addWidget(icon, 0, 0, 1, 1)
+
+    def mouseMoveEvent(self, e):
+        """ *
+        Override
+        """
+
+    def mousePressEvent(self, e):
+        """ *
+        Override
+        """
 
     def mouseReleaseEvent(self, e):
         """ *
